convert file_server to TypeScript
