+++ packages/https-proxy/lib/index.ts
@@ -2,3 +2,5 @@ import { create as createProxy, reset as resetProxy } from './proxy'
 import { CA } from './ca'
 
 export { createProxy, resetProxy, CA }
+
+export type { Server } from './server'
